Reogranize assignments storage tests
